[AC-7965] Apply format fix to reservation cancelation notification
