+++ web/impact/impact/tests/test_cancel_office_hour_reservation.py
@@ -1,3 +1,4 @@
+from django.core import mail
 from django.urls import reverse
 
 from accelerator.tests.factories import (
@@ -10,11 +11,11 @@ from ..v1.views import (
     NO_SUCH_RESERVATION,
     NO_SUCH_OFFICE_HOUR,
 )
-
 from ..v1.views.cancel_office_hour_reservation_view import (
     FAIL_HEADER,
     SUCCESS_HEADER,
 )
+from ..v1.views.utils import localized_office_hour_start_time
 
 
 class TestCancelOfficeHourReservationView(APITestCase):
@@ -126,6 +127,13 @@ class TestCancelOfficeHourReservationView(APITestCase):
         office_hour = MentorProgramOfficeHourFactory(finalist=None)
         self._submit_cancellation(office_hour, user=self.make_user())
         self.assert_not_notified(office_hour.mentor)
+
+    def test_correct_date_format_in_notification_email(self):
+        office_hour = MentorProgramOfficeHourFactory()
+        self._submit_cancellation(office_hour, user=office_hour.finalist)
+        email = mail.outbox[0]
+        localized_time = localized_office_hour_start_time(office_hour)
+        self.assertIn(localized_time.strftime("%I:%M"), email.body)
 
     def _submit_cancellation(self, office_hour, user, message=""):
         if office_hour is None:
