+++ web/impact/templates/emails/cancel_office_hour_reservation_email_to_finalist.html
@@ -1,8 +1,7 @@
 Hi {{ recipient.first_name }},
 
-Your upcoming office hour session at {{ office_hour_date_time | date:"h:m" }} on
-{{ office_hour_date_time | date }} with {{ counterpart.full_name }} was cancelled by
-{{ cancelling_party }}.
+Your upcoming office hour session at {{ start_time }} on {{ date }} with
+{{ counterpart.full_name }} was cancelled by {{ cancelling_party }}.
 To schedule a new time with this or any available mentor, visit
 https://accelerate.masschallenge.org/newofficehours
