+++ web/impact/templates/emails/cancel_office_hour_reservation_email_to_mentor.html
@@ -1,9 +1,8 @@
 Hi {{ recipient.first_name }},
 
-Your upcoming office hour session at {{ office_hour_date_time|date:"h:m" }} on
-{{ office_hour_date_time | date }} with {{ counterpart.full_name }} was cancelled
-by {{ cancelling_party }}. This office hour session is now available for another
-startup to reserve.
+Your upcoming office hour session at {{ start_time }} on {{ date }} with
+{{ counterpart.full_name }} was cancelled by {{ cancelling_party }}.
+This office hour session is now available for another startup to reserve.
 You may make changes to your upcoming office hours here:
 https://accelerate.masschallenge.org/newofficehours
